requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
pymilvus
langchain_google_genai
